refactor: decompose llm.rb god object into focused modules
lib/legion/llm.rb: 1133 lines → 155 lines thin facade
- start/shutdown/started?/settings/agent remain here
- chat/ask/chat_direct delegate to Inference module
- embed/embed_direct/embed_batch delegate to Call::Embeddings
- structured/structured_direct delegate to Call::StructuredOutput
- can_embed?/embedding_* delegate to Discovery module

lib/legion/llm/inference.rb: receives all chat/ask/escalation logic
- chat, ask, chat_direct (public module_function entries)
- dispatch_chat, chat_single, chat_with_escalation (routing logic)
- escalation/health/response guard/cache helpers
- diagnostic debug logging on every entry/exit/branch

lib/legion/llm/discovery.rb: receives all embedding discovery logic
- run (Ollama/System refresh at startup)
- detect_embedding_capability + fallback chain building
- reset! for clean shutdown
- @can_embed, @embedding_provider, @embedding_model state

lib/legion/llm/config.rb: receives set_defaults/auto_configure_defaults

lib/legion/llm/call/providers.rb: receives setup, resolve_llm_secrets,
auto_register_providers, inject_anthropic_cache_control!
