+++ lib/legion/llm/call/providers.rb
@@ -1,21 +1,51 @@
 # frozen_string_literal: true
 
+require 'legion/logging/helper'
+
 module Legion
   module LLM
     module Call
       module Providers
-        include Legion::Logging::Helper
+        extend Legion::Logging::Helper
+
+        module_function
+
+        def setup
+          log.debug '[llm][providers] setup.enter'
+          resolve_llm_secrets
+          configure_providers
+          verify_providers
+          auto_register_providers
+          log.debug '[llm][providers] setup.exit'
+        rescue StandardError => e
+          handle_exception(e, level: :error, operation: 'llm.providers.setup')
+          raise
+        end
+
+        def resolve_llm_secrets
+          log.debug '[llm][providers] resolve_llm_secrets.enter'
+          return unless defined?(Legion::Settings::Resolver)
+
+          Legion::Settings::Resolver.resolve_secrets!(Legion::LLM.settings)
+          log.debug '[llm][providers] resolve_llm_secrets.exit'
+        rescue StandardError => e
+          handle_exception(e, level: :warn, operation: 'llm.providers.resolve_llm_secrets')
+        end
 
         def configure_providers
+          log.debug '[llm][providers] configure_providers.enter'
           auto_enable_from_resolved_credentials
           Legion::LLM.settings[:providers].each do |provider, config|
             next unless config[:enabled]
 
+            log.debug "[llm][providers] configure_providers applying provider=#{provider}"
             apply_provider_config(provider, config)
           end
+          log.debug '[llm][providers] configure_providers.exit'
         end
 
         def auto_enable_from_resolved_credentials
+          log.debug '[llm][providers] auto_enable_from_resolved_credentials.enter'
           Legion::LLM.settings[:providers].each do |provider, config|
             next if config[:enabled]
 
@@ -35,7 +65,7 @@ module Legion
             next unless has_creds
 
             config[:enabled] = true
-            log.info "Auto-enabled #{provider} provider (credentials found)"
+            log.info "[llm][providers] auto-enabled provider=#{provider} reason=credentials_found"
           end
         end
 
@@ -45,6 +75,7 @@ module Legion
           host_part = url.gsub(%r{^https?://}, '').split(':')
           addr = host_part[0]
           port = (host_part[1] || '11434').to_i
+          log.debug "[llm][providers] ollama_running? addr=#{addr} port=#{port}"
           Socket.tcp(addr, port, connect_timeout: 1).close
           true
         rescue StandardError => e
@@ -54,20 +85,14 @@ module Legion
 
         def apply_provider_config(provider, config)
           case provider
-          when :bedrock
-            configure_bedrock(config)
-          when :anthropic
-            configure_anthropic(config)
-          when :openai
-            configure_openai(config)
-          when :gemini
-            configure_gemini(config)
-          when :azure
-            configure_azure(config)
-          when :ollama
-            configure_ollama(config)
+          when :bedrock   then configure_bedrock(config)
+          when :anthropic then configure_anthropic(config)
+          when :openai    then configure_openai(config)
+          when :gemini    then configure_gemini(config)
+          when :azure     then configure_azure(config)
+          when :ollama    then configure_ollama(config)
           else
-            log.warn "Unknown LLM provider: #{provider}"
+            log.warn "[llm][providers] unknown provider=#{provider}"
           end
         end
 
@@ -103,7 +128,7 @@ module Legion
           end
 
           auth_mode = has_bearer ? 'bearer token' : 'SigV4'
-          log.info "Configured Bedrock provider (#{config[:region]}, #{auth_mode})"
+          log.info "[llm][providers] configured bedrock region=#{config[:region]} auth=#{auth_mode}"
         end
 
         def configure_anthropic(config)
@@ -114,7 +139,7 @@ module Legion
             c.anthropic_api_key = api_key
             c.anthropic_api_base = config[:base_url] if config[:base_url]
           end
-          log.info "Configured Anthropic provider#{" (#{config[:base_url]})" if config[:base_url]}"
+          log.info "[llm][providers] configured anthropic base_url=#{config[:base_url].inspect}"
         end
 
         def configure_openai(config)
@@ -125,7 +150,7 @@ module Legion
             c.openai_api_key = api_key
             c.openai_api_base = config[:base_url] if config[:base_url]
           end
-          log.info "Configured OpenAI provider#{" (#{config[:base_url]})" if config[:base_url]}"
+          log.info "[llm][providers] configured openai base_url=#{config[:base_url].inspect}"
         end
 
         def configure_gemini(config)
@@ -136,7 +161,7 @@ module Legion
             c.gemini_api_key = api_key
             c.gemini_api_base = config[:base_url] if config[:base_url]
           end
-          log.info "Configured Gemini provider#{" (#{config[:base_url]})" if config[:base_url]}"
+          log.info "[llm][providers] configured gemini base_url=#{config[:base_url].inspect}"
         end
 
         def configure_azure(config)
@@ -150,19 +175,20 @@ module Legion
             c.azure_api_key = api_key if api_key
             c.azure_ai_auth_token = auth_token if auth_token
           end
-          log.info "Configured Azure AI Foundry provider (#{api_base})"
+          log.info "[llm][providers] configured azure api_base=#{api_base}"
         end
 
         def configure_ollama(config)
           RubyLLM.configure do |c|
             c.ollama_api_base = config[:base_url] if config[:base_url]
           end
-          log.info "Configured Ollama provider (#{config[:base_url]})"
+          log.info "[llm][providers] configured ollama base_url=#{config[:base_url].inspect}"
         end
 
         SAAS_PROVIDERS = %i[bedrock anthropic openai gemini azure].freeze
 
         def verify_providers
+          log.debug '[llm][providers] verify_providers.enter'
           Legion::LLM.settings[:providers].each do |provider, config|
             next unless config[:enabled]
             next unless SAAS_PROVIDERS.include?(provider)
@@ -177,16 +203,15 @@ module Legion
 
           enabled = Legion::LLM.settings[:providers].select { |_, c| c.is_a?(Hash) && c[:enabled] }
           if enabled.empty?
-            log.error 'No LLM providers available — all providers failed health checks or are disabled. ' \
-                      'LLM features (chat, inference, embeddings) will not work. ' \
-                      'Check API keys, network connectivity, and provider configuration.'
+            log.error '[llm][providers] no providers available — all failed health checks or disabled'
           else
             names = enabled.map { |name, c| "#{name}/#{c[:default_model] || 'auto'}" }
-            log.info "LLM providers available: #{names.join(', ')}"
+            log.info "[llm][providers] available providers=#{names.join(', ')}"
           end
         end
 
         def recover_with_alternative_credentials
+          log.debug '[llm][providers] recover_with_alternative_credentials.enter'
           recover_openai_with_codex
         end
 
@@ -197,7 +222,7 @@ module Legion
           token = Call::CodexConfigLoader.read_token
           return unless token
 
-          log.info 'OpenAI disabled — trying Codex auth token as fallback'
+          log.info '[llm][providers] openai disabled — trying codex auth token as fallback'
           openai_config[:api_key] = token
           configure_openai(openai_config)
           openai_config[:enabled] = true
@@ -210,15 +235,58 @@ module Legion
           start_time = Time.now
           RubyLLM.chat(model: model, provider: provider).ask('Respond with only the word: pong')
           elapsed = ((Time.now - start_time) * 1000).round
-          log.info "Health check #{provider}/#{model}: OK (#{elapsed}ms)"
+          log.info "[llm][providers] health_check ok provider=#{provider} model=#{model} elapsed_ms=#{elapsed}"
         rescue RubyLLM::ModelNotFoundError => e
           handle_exception(e, level: :warn, operation: 'llm.providers.verify_single_provider', provider: provider, model: model)
         rescue RubyLLM::ForbiddenError => e
           handle_exception(e, level: :debug, operation: 'llm.providers.verify_single_provider', provider: provider, model: model)
         rescue StandardError => e
-          log.warn "LLM provider (#{provider}) not available, #{e.class}"
+          log.warn "[llm][providers] provider unavailable provider=#{provider} error=#{e.class}"
           handle_exception(e, level: :debug, operation: 'llm.providers.verify_single_provider', provider: provider, model: model)
-          # config[:enabled] = false
+        end
+
+        def auto_register_providers
+          log.debug '[llm][providers] auto_register_providers.enter'
+          try_register_native_provider(:claude, 'Legion::Extensions::Claude', 'Legion::Extensions::Claude::Runners::Messages') do |klass|
+            Call::Registry.register(:claude, klass)
+            Call::Registry.register(:anthropic, klass)
+          end
+          try_register_native_provider(:bedrock, 'Legion::Extensions::Bedrock', 'Legion::Extensions::Bedrock::Runners::Converse') do |klass|
+            Call::Registry.register(:bedrock, klass)
+          end
+          try_register_native_provider(:openai, 'Legion::Extensions::Openai', 'Legion::Extensions::Openai::Runners::Chat') do |klass|
+            Call::Registry.register(:openai, klass)
+          end
+          try_register_native_provider(:gemini, 'Legion::Extensions::Gemini', 'Legion::Extensions::Gemini::Runners::Generate') do |klass|
+            Call::Registry.register(:gemini, klass)
+          end
+
+          registered = Call::Registry.available
+          if registered.any?
+            log.info "[llm][providers] native registry registered=#{registered.join(', ')}"
+          else
+            log.debug '[llm][providers] no native lex-* providers registered (ruby_llm mode)'
+          end
+        rescue StandardError => e
+          handle_exception(e, level: :warn, operation: 'llm.providers.auto_register')
+        end
+
+        def inject_anthropic_cache_control!(opts, provider)
+          resolved_provider = (provider || Legion::LLM.settings[:default_provider])&.to_sym
+          return unless resolved_provider == :anthropic
+
+          caching_settings = Legion::LLM.settings[:prompt_caching] || {}
+          return unless caching_settings[:enabled] != false
+
+          min_tokens = caching_settings[:min_tokens] || 1024
+          instructions = opts[:instructions]
+          return unless instructions.is_a?(String) && instructions.length > min_tokens
+
+          log.debug "[llm][providers] inject_anthropic_cache_control provider=#{resolved_provider} length=#{instructions.length}"
+          opts[:instructions] = {
+            content:       instructions,
+            cache_control: { type: 'ephemeral' }
+          }
         end
 
         def resolve_broker_credential(provider_name)
@@ -255,6 +323,15 @@ module Legion
         rescue StandardError => e
           handle_exception(e, level: :debug, operation: 'llm.providers.broker_credential_available', provider: provider)
           false
+        end
+
+        def try_register_native_provider(name, ext_const, runner_const)
+          log.debug "[llm][providers] try_register_native_provider name=#{name} ext=#{ext_const}"
+          return unless Object.const_defined?(ext_const, false) && Object.const_defined?(runner_const, false)
+
+          klass = Object.const_get(runner_const)
+          yield klass
+          log.debug "[llm][providers] registered native provider name=#{name}"
         end
       end
     end
